Move ParentData classes to ScopeModifiers file
