--- grid/src/commonMain/kotlin/com/cheonjaeung/compose/grid/SequentialGridParentData.kt
@@ -1,12 +1,0 @@
-package com.cheonjaeung.compose.grid
-
-import androidx.compose.ui.Alignment
-
-/**
- * A data class to contain modifier options of item at the specified position.
- */
-internal data class SequentialGridParentData(
-    var span: (GridItemSpanScope.() -> Int)? = null,
-    var alignment: Alignment? = null,
-    var mainAxisSizeFraction: Float? = null
-)
